Code of Lesson1:
Lookup the System.DateTime structure. This type contains a large amount
of functionality that allows a user to perform all sorts of interesting
operations with dates and times. Use System.DateTime along with
System.Console to implement a # program that does the following:

Ask the user for their birthday. It will probably easiest to ask year,
then month, then day rather than parsing a combined string reliably.
Calculate the age of the user.
Check to see if the age of the user is impossible. For example, if the
user is not yet born, output an error message. If the user claims to be
135 years old, then let them know that's not possible.

Task: Zigger-lesson1